+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-# Craft CMS 3 - Sms77
+# Craft CMS 3 - sms77
 
 ## v1.0.0 - 2021-04-19
 
 ### Added
 
-- Initial release of the Craft CMS 3 plugin for [Sms77](https://www.sms77.io)
+- Initial release of the Craft CMS 3 plugin for [sms77](https://www.sms77.io)